app/requirements.txt
fastapi
sqlalchemy
pydantic
uvicorn
psycopg2
python-jose
passlib
python-multipart